Dissolve versionmanager
Move to /src/lib: getCurrentDependencies, getInstalledPackages, upgradePackageData
