+++ src/lib/getPeerDependencies.ts
@@ -1,7 +1,7 @@
 import fs from 'fs'
 import jph from 'json-parse-helpfulerror'
-import * as vm from '../versionmanager'
 import { print } from '../logging'
+import getCurrentDependencies from './getCurrentDependencies'
 import { Index, Options, VersionDeclaration } from '../types'
 
 /** Get peer dependencies from installed packages */
@@ -13,7 +13,7 @@ function getPeerDependencies(current: Index<VersionDeclaration>, options: Option
     try {
       const pkgData = fs.readFileSync(path, 'utf-8')
       const pkg = jph.parse(pkgData)
-      peers = vm.getCurrentDependencies(pkg, { ...options, dep: 'peer' })
+      peers = getCurrentDependencies(pkg, { ...options, dep: 'peer' })
     }
     catch (e) {
       print(options, 'Could not read peer dependencies for package ' + pkgName + '. Is this package installed?', 'warn')
